# public/api/openapi.yaml
openapi: 3.1.0
info:
  title: NuCRM API
  description: |
    The NuCRM REST API provides programmatic access to CRM functionality including contacts,
    companies, deals, tasks, invoices, and more. All API requests must be authenticated
    and scoped to a specific tenant.

    ## Authentication
    All API endpoints require authentication via cookie-based session or Bearer token.
    Include the session cookie (`nucrm_session`) or `Authorization: Bearer <token>` header.

    ## Tenant Context
    Most endpoints operate within a tenant context. The tenant is determined by the
    authenticated user's session or the `X-Tenant-ID` header for super admin operations.

    ## Rate Limiting
    API requests are rate-limited to 100 requests per minute per user.

    ## Pagination
    List endpoints support pagination via `limit` and `offset` query parameters.
    Default: limit=50, offset=0. Maximum limit: 500.
  version: 2.0.0
  contact:
    name: NuCRM API Support
    email: api@nucrm.com
  license:
    name: Proprietary
servers:
  - url: https://api.nucrm.com/api/v2
    description: Production API
  - url: http://localhost:3000/api/v2
    description: Local development
...
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: nucrm_session
      description: Session cookie issued after login
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for API access
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: "Invalid request payload"
        code:
          type: string
          example: "VALIDATION_ERROR"
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
    Pagination:
      type: object
      properties:
        total:
          type: integer
          example: 150
        limit:
          type: integer
          example: 50
        offset:
          type: integer
          example: 0
        hasMore:
          type: boolean
          example: true
    Contact:
      type: object
      required: [firstName, lastName]
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        firstName:
          type: string
          maxLength: 100
          example: "John"
        lastName:
          type: string
          maxLength: 100
          example: "Doe"
        email:
          type: string
          format: email
          nullable: true
          example: "john@example.com"
        phone:
          type: string
          nullable: true
          example: "+1-555-0123"
        jobTitle:
          type: string
          nullable: true
          example: "CEO"
        companyId:
          type: string
          format: uuid
          nullable: true
        ownerId:
          type: string
          format: uuid
          nullable: true
        tags:
          type: array
          items:
            type: string
          example: ["vip", "enterprise"]
        customFields:
          type: object
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
...
  parameters:
    limit:
      name: limit
      in: query
      schema:
        type: integer
        default: 50
        maximum: 500
      description: Number of items to return
    offset:
      name: offset
      in: query
      schema:
        type: integer
        default: 0
      description: Number of items to skip
    search:
      name: search
      in: query
      schema:
        type: string
      description: Search term
    sort:
      name: sort
      in: query
      schema:
        type: string
        default: "createdAt:desc"
      description: Sort field and direction (field:asc|desc)
security:
  - SessionCookie: []
  - BearerAuth: []
paths:
  /contacts:
    get:
      tags: [Contacts]
      summary: List contacts
      description: Retrieve a paginated list of contacts with optional filtering
      parameters:
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
        - $ref: '#/components/parameters/search'
        - $ref: '#/components/parameters/sort'
        - name: companyId
          in: query
          schema:
            type: string
            format: uuid
          description: Filter by company
        - name: ownerId
          in: query
          schema:
            type: string
            format: uuid
          description: Filter by owner
        - name: tags
          in: query
          schema:
            type: array
            items:
              type: string
          description: Filter by tags
      responses:
        '200':
          description: List of contacts
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Contact'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden

    post:
      tags: [Contacts]
      summary: Create contact
      description: Create a new contact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Contact'
      responses:
        '201':
          description: Contact created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contact'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...